feat: work create 컬럼 2개에 대한 테스트 기능 구현
