+++ webapp/models/__init__.py
@@ -1,2 +1,8 @@
 # __init__.py
 from .sales import *
+from .work import Work
+from .worksize import WorkSize
+from .picture import Picture
+from .theme import Theme
+from .unit import Unit
+from .artist import Artist
