+++ webapp/urls.py
@@ -1,6 +1,7 @@
 from django.urls import path
 
 from webapp.views import views
+from webapp.views import work
 
 urlpatterns = [
     path("organizations/", views.ReadOrganizationAPI.as_view(), name="Read-Organization"),
@@ -10,8 +11,8 @@ urlpatterns = [
         name="Create-Organization",
     ),
     path(
-        "work/create",
-        views.CreateWorkAPI.as_view(),
+        "work/create/seoul",
+        work.CreateSeoulWorkAPI.as_view(),
         name="Create-Work",
     ),
 ]
